restructure code base
